fix(uplc): resolve clang-tidy findings in the UPLC sources
Clear the clang-tidy warnings the analysis build treats as errors: widen the find-first-set-bit arithmetic in the target type, merge identical switch and conditional branches into their default arms, drop top-level const from a forward-declared parameter, and align a forward declaration's parameter names with the definition.

+++ lib/src/uplc/arena/uplc_arena.c
@@ -35,7 +35,7 @@
  * \brief Default block payload size, used when the caller passes 0 to
  *        \ref cardano_uplc_arena_new.
  */
-static const size_t ARENA_DEFAULT_BLOCK_SIZE = 64U * 1024U;
+static const size_t ARENA_DEFAULT_BLOCK_SIZE = (size_t)64U * 1024U;
 
 /**
  * \brief Upper bound on the total payload bytes the arena will serve.
